+++ fe/src/components/IssueTable/IssueTableHeader/OpenAndCloseFilter/index.tsx
@@ -5,7 +5,7 @@ import { useNavigate } from 'react-router-dom';
 import Container from '@components/Container';
 import IconTextBox from '@components/IconTextBox';
 import { fontWeight } from '@constants/fonts';
-import { handleUrlQuery } from '@util/queryParser';
+import { makeUrlQuery } from '@util/queryParser';
 
 export default function OpenAndCloseFilter({
   clickedStatusCnt,
@@ -18,7 +18,7 @@ export default function OpenAndCloseFilter({
   const searchParams = new URLSearchParams(window.location.search);
   const clickedStatus = searchParams.get('status');
   const handleClickStatusTab = (status) => {
-    const queryString = handleUrlQuery('set', 'status', status);
+    const queryString = makeUrlQuery('set', 'status', status);
     navigate(`/?${queryString}`);
   };
