+++ fe/src/server/handlers.ts
@@ -1,7 +1,7 @@
 import { rest } from 'msw';
 
 import issues from '@server/dummyData/issues';
-import { IssueType } from '@type/types';
+import { filterIssues } from '@server/filterUtil';
 
 interface IUser {
   member_id: number;
@@ -46,65 +46,9 @@ const postLogin: Parameters<typeof rest.get>[1] = async (req, res, ctx) => {
   return res(ctx.status(401));
 };
 
-
 const getIssues: Parameters<typeof rest.get>[1] = (req, res, ctx) => {
-  const filteredIssues = filterIssues(req.url.search);
+  const filteredIssues = filterIssues(req.url.search, issues);
   return res(ctx.status(200), ctx.json(filteredIssues));
-};
-
-const filterIssues = (queryString) => {
-  const searchParams = new URLSearchParams(queryString);
-  const queryKeyList = ['writer', 'label', 'milestone', 'assignee'];
-  const filteredByQueries = queryKeyList.reduce(
-    (tempIssues, queryKey) =>
-      filterByQuery(queryKey, searchParams.get(queryKey), tempIssues),
-    issues,
-  );
-  const filteredByStatus =
-    searchParams.get('status') === 'closed'
-      ? filterByStatus('closed', filteredByQueries)
-      : filterByStatus('open', filteredByQueries);
-  return filteredByStatus;
-};
-
-const filterByQuery = (
-  queryKey: string,
-  queryValue: string,
-  originalIssues: IssueType[],
-) => {
-  if (queryValue === null) return originalIssues;
-  switch (queryKey) {
-    case 'writer':
-      return originalIssues.filter((issue) => issue.writer === queryValue);
-    case 'label':
-      return filterByLabel(Number(queryValue), originalIssues);
-    case 'milestone':
-      return originalIssues.filter(
-        (issue) => issue.milestone.id === Number(queryValue),
-      );
-    case 'assignee':
-      return originalIssues.filter((issue) =>
-        issue.assignee.includes(queryValue),
-      );
-    default:
-      return originalIssues;
-  }
-};
-
-const filterByLabel = (targetLabelId: number, originalIssues: IssueType[]) =>
-  originalIssues.filter((issue) =>
-    issue.labels.some((label) => label.id === targetLabelId),
-  );
-
-const filterByStatus = (
-  target: 'open' | 'closed',
-  originalIssues: IssueType[],
-) => {
-  const filtered = originalIssues.filter((issue) => issue.status === target);
-  return {
-    issues: filtered,
-    oppositeStatusCnt: originalIssues.length - filtered.length,
-  };
 };
 
 export default function handlers() {
